requirements.txt
sentence_transformers
fastapi
python-gitlab